pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "wechat-agent"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
keywords = []
authors = [
  { name = "jerry", email = "826935261@qq.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "wxauto~=39.1.16",
  "langchain~=0.3.27",
  "langchain-community~=0.3.27",
  "langchain-text-splitters~=0.3.11",
  "langchain-core~=0.3.76",
  "langchain-experimental~=0.3.4",
  "flask~=3.1.2",
  "pyjwt~=2.10.1",

]

[project.urls]
Documentation = "https://github.com/jerry/wechat-agent#readme"
Issues = "https://github.com/jerry/wechat-agent/issues"
Source = "https://github.com/jerry/wechat-agent"

[tool.hatch.version]
path = "src/wechat_agent/__about__.py"

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/wechat_agent tests}"

[tool.coverage.run]
source_pkgs = ["wechat_agent", "tests"]
branch = true
parallel = true
omit = [
  "src/wechat_agent/__about__.py",
]

[tool.coverage.paths]
wechat_agent = ["src/wechat_agent", "*/wechat-agent/src/wechat_agent"]
tests = ["tests", "*/wechat-agent/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
